# apps/api/src/resources/swagger.yaml
...
openapi: 3.0.0
info:
  title: PCGL DACO API
  version: 1.0.0
paths:
...
  /collaborators/create:
    post:
      tags:
        - Collaborators
      summary: Create new Collaborators for an application.
      requestBody:
        description: The body containing information about the requesting user.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicationId:
                  type: number
                  description: ID of the related Application record that new Collaborators are associated with
                userId:
                  type: string
                  description: Temporary, Applicant's User Id, required for authorization
                collaborators:
                  $ref: '#/components/requests/CreateCollaborators'
      responses:
        '500':
          description: 'Error creating collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Error creating Collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
        '201':
          description: 'Collaborators created'
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/responses/CollaboratorsListResponse'
  /collaborators/{applicationId}:
    get:
      tags:
        - Collaborators
      summary: List all Collaborators for an application.
      parameters:
        - in: path
          name: applicationId
          schema:
            type: number
          description: Id of application to find related collaborators
          required: true
      responses:
        '201':
          description: 'Collaborators retrieved'
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/responses/CollaboratorsListResponse'
        '500':
          description: 'Error listing collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Error listing Collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
  /collaborators/update:
    post:
      tags:
        - Collaborators
      summary: Edit existing Collaborator records
      requestBody:
        description: The body containing information about the requesting user.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applicationId:
                  type: number
                  description: ID of the related Application record that Collaborator records are associated with
                userId:
                  type: string
                  description: Temporary, Applicant's User Id, required for authorization
                collaboratorUpdates:
                  $ref: '#/components/requests/UpdatedCollaborators'
      responses:
        '201':
          description: 'Collaborators updated'
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/responses/CollaboratorsListResponse'
        '500':
          description: 'Error updating collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Error updating Collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ClientErrors'
  /collaborators/{applicationId}/{collaboratorId}:
    delete:
      tags:
        - Collaborators
      summary: Remove collaborator from an application.
      parameters:
        - in: path
          name: applicationId
          schema:
            type: integer
          required: true
          description: Numeric ID of the relevant application that the collaborator will be removed form.
        - in: path
          name: collaboratorId
          schema:
            type: integer
          required: true
          description: Numeric ID of the Collaborator to delete
      responses:
        '201':
          description: 'Collaborator deleted'
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/responses/CollaboratorsListResponse'
        '500':
          description: 'Error deleting collaborators'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Error deleting Collaborators. Request validation failed.'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/RequestValidationError'
...
  /file/ethics/{applicationId}:
    post:
      tags:
        - File
      summary: Upload new file or replace old file
      parameters:
        - in: path
          name: applicationId
          schema:
            type: integer
          required: true
          description: Numeric ID of the associated application upload
      consumes:
        - multipart/form-data
      requestBody:
        description: |
          Files must fit a certain set of properties to be accepted by our server. The file cannot be more than `5mb`, has to be of type `'pdf' 'doc' 'docx' ` and only `one` file can be uploaded.
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 'File uploaded'
          content:
            application/json:
              schema:
                type: object
        '500':
          description: 'Error uploading file to database'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Error uploading file. File must be less than 5mb and only 1 file can be uploaded.'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/RequestValidationError'
                  - $ref: '#/components/responses/ClientErrors'
  /file/{fileId}:
    delete:
      tags:
        - File
      summary: Delete uploaded by file id
      parameters:
        - in: path
          name: fileId
          schema:
            type: integer
          required: true
          description: Numeric ID of the associated fileId upload
      responses:
        '200':
          description: 'File deleted'
          content:
            application/json:
              schema:
                type: object
        '500':
          description: 'Error deleting file in database'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerErrors'
        '400':
          description: 'Invalid fileId.'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/RequestValidationError'
                  - $ref: '#/components/responses/ClientErrors'
...
components:
  schemas:
...
    Collaborators:
      type: object
      properties:
        id:
          type: number
          description: The unique ID of the collaborator in the database.
        application_id:
          type: number
          description: The Application ID associated with the Collaborator.
        first_name:
          type: string
          description: The collaborator's first name
        middle_name:
          type: string
          description: The collaborator's middle name
        last_name:
          type: string
          description: The collaborator's last name
        suffix:
          type: string
          description: The collaborator's additional name suffix
        position_title:
          type: string
          description: The collaborator's job title
        institutional_email:
          type: string
          description: The collaborator's contact email
...
  requests:
...
    CreateCollaborators:
      type: array
      description: Array of new Collaborators to create
      items:
        $ref: '#/components/requests/NewCollaborator'
    NewCollaborator:
      type: object
      required:
        - collaboratorFirstName
        - collaboratorLastName
        - collaboratorPositionTitle
        - collaboratorInstitutionalEmail
      properties:
        collaboratorFirstName:
          type: string
          description: The collaborator's first name
        collaboratorMiddleName:
          type: string
          description: The collaborator's middle name
        collaboratorLastName:
          type: string
          description: The collaborator's last name
        collaboratorSuffix:
          type: string
          description: The collaborator's additional name suffix
        collaboratorPositionTitle:
          type: string
          description: The collaborator's job title
        collaboratorInstitutionalEmail:
          type: string
          description: The collaborator's contact email
    UpdatedCollaborators:
      type: object
      required:
        - id
      properties:
        id:
          type: number
          description: ID of the Collaborator record to update
        collaboratorFirstName:
          type: string
          description: The collaborator's first name
        collaboratorMiddleName:
          type: string
          description: The collaborator's middle name
        collaboratorLastName:
          type: string
          description: The collaborator's last name
        collaboratorSuffix:
          type: string
          description: The collaborator's additional name suffix
        collaboratorPositionTitle:
          type: string
          description: The collaborator's job title
        collaboratorInstitutionalEmail:
          type: string
          description: The collaborator's contact email
...
  responses:
...
    ClientErrors:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: General description of error with parameters and context
    CollaboratorsListResponse:
      type: array
      description: Array of Collaborator records
      items:
        $ref: '#/components/schemas/Collaborators'
    ServerErrors:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: General description of error with parameters and context
        errors:
          description: Detailed error data thrown by code failure
          type: string
...
    RequestValidationError:
      type: object
      required:
        - message
        - error
        - details
      properties:
        error:
          type: string
          description: Error code, usually "RequestValidationError"
        message:
          type: string
          description: General description of error with parameters and context
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                description: The validation error type
              expected:
                type: string
                description: The expected property.
              received:
                type: string
                description: The property that was received in the request
              path:
                type: array
                items:
                  type: string
                  description: The path of the object where validation failed.
              message:
                type: string
                description: A plain text error message describing why the validation failed.